Fixed some bugs & preparing for more automatic exploration

+++ synthesizable/include/core.h
@@ -55,7 +55,7 @@ struct MemtoWB
 
 void doStep(ac_int<32, false> pc, ac_int<32, true> ins_memory[8192], unsigned int dm[8192], bool &exit
 #ifndef __SYNTHESIS__
-, int cycles
+, uint64_t cycles
 #endif
 );
